Mdx refact
