--- mdx/custom-mapping.tsx
@@ -1,11 +1,0 @@
-import { CommandItem } from "~/components/CommandItem";
-import { CustomH2 } from "~/components/CustomH2";
-import { CustomH3 } from "~/components/CustomH3";
-import { Link } from "~/components/Link";
-
-export const components = {
-  a: Link,
-  C: CommandItem,
-  h2: CustomH2,
-  h3: CustomH3,
-};
